added types/ resolving comments
