+++ producers/pipsafety/main.go
@@ -1,39 +1,17 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 
 	v1 "api/proto/v1"
 
+	"producers/pipsafety/types"
+
 	"github.com/thought-machine/dracon/producers"
 )
 
-type SafetyIssue struct {
-	Name              string
-	VersionConstraint string
-	CurrentVersion    string
-	Description       string
-}
-
-//read semi-unstructured safety json into struct
-func (i *SafetyIssue) UnmarshalJSON(data []byte) error {
-
-	var v []interface{}
-	if err := json.Unmarshal(data, &v); err != nil {
-		return err
-	}
-
-	i.Name, _ = v[0].(string)
-	i.VersionConstraint, _ = v[1].(string)
-	i.CurrentVersion = v[2].(string)
-	i.Description = v[3].(string)
-
-	return nil
-}
-
-func parseIssues(out []SafetyIssue) []*v1.Issue {
+func parseIssues(out []types.SafetyIssue) []*v1.Issue {
 	issues := []*v1.Issue{}
 	for _, r := range out {
 		issues = append(issues, &v1.Issue{
@@ -53,7 +31,7 @@ func main() {
 	if err := producers.ParseFlags(); err != nil {
 		log.Fatal(err)
 	}
-	issues := []SafetyIssue{}
+	issues := []types.SafetyIssue{}
 	producers.ParseInFileJSON(&issues)
 	if err := producers.WriteDraconOut(
 		"pipsafety",
