+++ producers/pipsafety/types/BUILD
@@ -1,0 +1,9 @@
+
+go_library(
+    name = "types",
+    srcs = [
+        "safety-issue.go",
+    ],
+    visibility = ["//producers/pipsafety/..."]
+
+)
